Move globals to separate file
