+++ src/utils.ts
@@ -1,18 +1,3 @@
-/// <reference lib="dom" />
-
 export function noop(): undefined {
   return undefined;
 }
-
-function getGlobals() {
-  if (typeof self !== 'undefined') {
-    return self;
-  } else if (typeof window !== 'undefined') {
-    return window;
-  } else if (typeof global !== 'undefined') {
-    return global;
-  }
-  return undefined;
-}
-
-export const globals = getGlobals();
